refactor: Phase 4a — model/loss registries, shared helpers, head consolidation (BPR)
- models/registry.py: @register_model + create_model(), 16 models registered
- losses/registry.py: @register_loss + create_loss(), 5 losses registered
- models/heads/category.py: all 8 category heads consolidated
- models/heads/next.py: all 7 next heads consolidated
- models/components/positional.py: shared PositionalEncoding
- training/helpers.py: compute_class_weights, setup_optimizer, setup_scheduler
- training/evaluate.py: collect_predictions, build_report
- Original model files become thin re-export shims
- Category + next runners wired to registries and shared helpers
