+++ src/losses/__init__.py
@@ -1,0 +1,1 @@
+"""Loss function registry — created in Phase 4a."""
