+++ src/model/next/next_head.py
@@ -1,113 +1,3 @@
-import math
-import torch
-import torch.nn as nn
-
-
-class PositionalEncoding(nn.Module):
-    def __init__(self, embed_dim, max_seq_length=5000, dropout=0.1):
-        """
-        Args:
-            embed_dim: The dimensionality of the embedding space.
-            max_seq_length: The maximum sequence length to handle.
-            dropout: Dropout rate applied to positional encodings.
-        """
-        super().__init__()
-        self.dropout = nn.Dropout(p=dropout)
-
-        # Precompute the positional encodings once
-        pe = torch.zeros(max_seq_length, embed_dim)
-        pos = torch.arange(0, max_seq_length, dtype=torch.float).unsqueeze(1)
-        div_term = torch.exp(
-            torch.arange(0, embed_dim, 2).float() * (-math.log(10000.0) / embed_dim)
-        )
-
-        pe[:, 0::2] = torch.sin(pos * div_term)
-        pe[:, 1::2] = torch.cos(pos * div_term)
-        pe = pe.unsqueeze(0)  # shape (1, max_seq_length, embed_dim)
-
-        self.register_buffer('positional_encoding', pe)
-        self.eps = 1e-12
-
-    def forward(self, x, padding_mask=None):
-        """
-        x: (batch_size, seq_length, embed_dim)
-        padding_mask: (batch_size, seq_length), True for padding positions
-        """
-        seq_length = x.size(1)
-        pos_enc = self.positional_encoding[:, :seq_length, :]
-
-        if padding_mask is not None:
-            # zero out both x and pos_enc where padding_mask == True
-            mask = (~padding_mask).unsqueeze(-1).float() + self.eps
-            x = x * mask
-            pos_enc = pos_enc * mask
-
-        x = x + pos_enc
-        return self.dropout(x)
-
-
-class NextHeadSingle(nn.Module):
-    def __init__(self, embed_dim, num_classes, num_heads, seq_length, num_layers, dropout=0.2):
-        super().__init__()
-        self.embed_dim = embed_dim
-        self.seq_length = seq_length
-
-        # Learned positional embeddings (better than sinusoidal for short fixed-length sequences)
-        self.pos_embedding = nn.Parameter(torch.randn(1, seq_length, embed_dim) * 0.02)
-        self.dropout = nn.Dropout(dropout)
-
-        encoder_layer = nn.TransformerEncoderLayer(
-            d_model=embed_dim,
-            nhead=num_heads,
-            dim_feedforward=embed_dim * 4,
-            dropout=dropout,
-            batch_first=True,
-            norm_first=True
-        )
-        self.transformer_encoder = nn.TransformerEncoder(encoder_layer, num_layers)
-
-        # Project each time step down to a single logit for pooling
-        self.sequence_reduction = nn.Linear(embed_dim, 1)
-        self.layer_norm = nn.LayerNorm(embed_dim)
-        self.classifier = nn.Linear(embed_dim, num_classes)
-
-        # Temporal decay bias for attention pooling (recent visits weighted more)
-        # Initialized as linear ramp: oldest=-2.0, newest=0.0 (learnable)
-        self.temporal_bias = nn.Parameter(torch.linspace(-2.0, 0.0, seq_length))
-
-    def forward(self, x, return_attention=False):
-        """
-        x: (batch_size, seq_length, embed_dim)
-        return_attention: If True, return (logits, attn_weights) where attn_weights is (batch, seq)
-        """
-        batch_size, seq_length, _ = x.size()
-        device = x.device
-
-        padding_mask = (x.abs().sum(dim=-1) == 0)  # (batch, seq), True for pad
-
-        # Add learned positional embeddings and mask padding
-        mask = (~padding_mask).unsqueeze(-1).float()
-        x = (x + self.pos_embedding[:, :seq_length, :]) * mask
-        x = self.dropout(x)
-
-        x = self.transformer_encoder(
-            x,
-            src_key_padding_mask=padding_mask
-        )
-
-        x = self.layer_norm(x)
-
-        attn_logits = self.sequence_reduction(x).squeeze(-1)
-        attn_logits = attn_logits + self.temporal_bias[:seq_length].unsqueeze(0)
-        attn_logits = attn_logits.masked_fill(padding_mask, float('-inf'))
-
-        attn_weights = torch.softmax(attn_logits, dim=1)
-        attn_weights = torch.nan_to_num(attn_weights, nan=0.0)  # (batch, seq)
-
-        pooled = torch.sum(x * attn_weights.unsqueeze(-1), dim=1)  # (batch, embed_dim)
-
-        out = self.classifier(pooled)
-
-        if return_attention:
-            return out, attn_weights
-        return out
+"""Shim — canonical location is models.heads.next (Phase 4a)."""
+from models.components.positional import PositionalEncoding  # noqa: F401
+from models.heads.next import NextHeadSingle  # noqa: F401
