+++ src/models/__init__.py
@@ -1,0 +1,5 @@
+"""Consolidated model definitions and registry.
+
+Created in Phase 4a. Canonical location for all model classes.
+Old src/model/ files become thin re-export shims after consolidation.
+"""
