+++ src/training/__init__.py
@@ -1,0 +1,1 @@
+"""Shared training and evaluation primitives — created in Phase 4a."""
